move resize function to a better place
